LinkedIn-AI-News-Generator/requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.4.0
feedparser>=6.0.0
GoogleNews>=1.6.0
python-dotenv>=0.20.0
lxml_html_clean>=0.1.0
newspaper3k>=0.2.8

# LangChain and AI dependencies
langchain>=0.1.0
langchain-huggingface>=0.0.10
langchain-community>=0.1.0
huggingface_hub>=0.16.0
transformers>=4.30.0
sentence_transformers>=2.2.0
tavily-python>=0.2.2

# Only needed for local interactive use (optional in VSCode)
ipywidgets>=8.0.0